writing to db works

+++ common/performance_model/performance_models/rob_performance_model/pics/dip.cc
@@ -21,22 +21,22 @@ void DipStack::add_base(SubsecondTime t)
     base += t;
 }
 
-UInt64 DipStack::get_base_cyc() const
+float DipStack::get_base_cyc() const
 {
     return static_cast<float>(base.getFS()) / static_cast<float>(Sim()->getDvfsManager()->getGlobalDomain()->getPeriod().getFS());
 }
 
-UInt64 DipStack::get_fe_stall_cyc() const
+float DipStack::get_fe_stall_cyc() const
 {
     return static_cast<float>(fe_stall.getFS()) / static_cast<float>(Sim()->getDvfsManager()->getGlobalDomain()->getPeriod().getFS());
 }
 
-UInt64 DipStack::get_be_stall_cyc() const
+float DipStack::get_be_stall_cyc() const
 {
     return static_cast<float>(be_stall.getFS()) / static_cast<float>(Sim()->getDvfsManager()->getGlobalDomain()->getPeriod().getFS());
 }
 
-UInt64 DipStack::get_mispred_cyc() const
+float DipStack::get_mispred_cyc() const
 {
     return static_cast<float>(mispred.getFS()) / static_cast<float>(Sim()->getDvfsManager()->getGlobalDomain()->getPeriod().getFS());
 }
